+++ src/main/java/com/guilhermeramos31/springbootjuniorcase/services/author/interfaces/AuthorService.java
@@ -1,12 +1,10 @@
-package com.guilhermeramos31.springbootjuniorcase.service.author.interfaces;
+package com.guilhermeramos31.springbootjuniorcase.services.author.interfaces;
 
 import com.guilhermeramos31.springbootjuniorcase.model.author.Author;
 import com.guilhermeramos31.springbootjuniorcase.model.author.dto.AuthorPaginationRequestDTO;
 import com.guilhermeramos31.springbootjuniorcase.model.author.dto.AuthorPaginationResponseDTO;
 import com.guilhermeramos31.springbootjuniorcase.model.author.dto.AuthorRequestDTO;
 import com.guilhermeramos31.springbootjuniorcase.model.author.dto.AuthorResponseDTO;
-
-import java.util.List;
 
 public interface AuthorService {
     AuthorResponseDTO save(AuthorRequestDTO authorRequestDTO);
