+++ src/main/java/com/guilhermeramos31/springbootjuniorcase/services/category/interfaces/CategoryService.java
@@ -1,11 +1,10 @@
-package com.guilhermeramos31.springbootjuniorcase.service.category.interfaces;
+package com.guilhermeramos31.springbootjuniorcase.services.category.interfaces;
 
 import com.guilhermeramos31.springbootjuniorcase.model.category.Category;
 import com.guilhermeramos31.springbootjuniorcase.model.category.dto.CategoryPaginationRequestDTO;
 import com.guilhermeramos31.springbootjuniorcase.model.category.dto.CategoryPaginationResponseDTO;
 import com.guilhermeramos31.springbootjuniorcase.model.category.dto.CategoryRequestDTO;
 import com.guilhermeramos31.springbootjuniorcase.model.category.dto.CategoryResponseDTO;
-import jakarta.validation.constraints.NotNull;
 
 public interface CategoryService {
     CategoryResponseDTO create(CategoryRequestDTO categoryRequestDTO);
